+++ ionoscloud/resource_dns_zone_test.go
@@ -11,7 +11,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 	dns "github.com/ionos-cloud/sdk-go-dns"
-	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services"
+
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/bundleclient"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/utils/constant"
 )
 
@@ -92,7 +93,7 @@ func TestAccDNSZone(t *testing.T) {
 }
 
 func testAccDNSZoneDestroyCheck(s *terraform.State) error {
-	client := testAccProvider.Meta().(services.SdkBundle).DNSClient
+	client := testAccProvider.Meta().(bundleclient.SdkBundle).DNSClient
 	ctx, cancel := context.WithTimeout(context.Background(), *resourceDefaultTimeouts.Default)
 	defer cancel()
 
@@ -115,7 +116,7 @@ func testAccDNSZoneDestroyCheck(s *terraform.State) error {
 
 func testAccDNSZoneExistenceCheck(path string, zone *dns.ZoneRead) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
-		client := testAccProvider.Meta().(services.SdkBundle).DNSClient
+		client := testAccProvider.Meta().(bundleclient.SdkBundle).DNSClient
 		rs, ok := s.RootModule().Resources[path]
 
 		if !ok {