+++ services/autoscaling/client.go
@@ -8,28 +8,30 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/meta"
 	autoscaling "github.com/ionos-cloud/sdk-go-bundle/products/vmautoscaling/v2"
 	"github.com/ionos-cloud/sdk-go-bundle/shared"
+	"github.com/ionos-cloud/sdk-go-bundle/shared/fileconfiguration"
 
-	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/utils"
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/clientoptions"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/utils/constant"
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/utils/loadedconfig"
 )
 
+// Client - wrapper over sdk client
 type Client struct {
 	sdkClient *autoscaling.APIClient
 }
 
-func NewClient(username, password, token, url, version, terraformVersion string, insecure bool) *Client {
-	newAutoscalingConfig := shared.NewConfiguration(username, password, token, url)
+// NewClient - returns a new autoscaling client
+func NewClient(clientOptions clientoptions.TerraformClientOptions, fileConfig *fileconfiguration.FileConfig) *Client {
+	loadedconfig.SetGlobalClientOptionsFromFileConfig(&clientOptions, fileConfig, fileconfiguration.Autoscaling)
+	config := shared.NewConfiguration(clientOptions.Credentials.Username, clientOptions.Credentials.Password, clientOptions.Credentials.Token, clientOptions.Endpoint)
+	config.UserAgent = fmt.Sprintf(
+		"terraform-provider/_ionos-cloud-sdk-go-vm-autoscaling/%s_hashicorp-terraform/%s_terraform-plugin-sdk/%s_os/%s_arch/%s",
+		autoscaling.Version, clientOptions.TerraformVersion, meta.SDKVersionString(), runtime.GOOS, runtime.GOARCH) //nolint:staticcheck
 
-	newAutoscalingConfig.MaxRetries = constant.MaxRetries
-	newAutoscalingConfig.MaxWaitTime = constant.MaxWaitTime
-	newAutoscalingConfig.UserAgent = fmt.Sprintf(
-		"terraform-provider/%s_ionos-cloud-sdk-go-vm-autoscaling/%s_hashicorp-terraform/%s_terraform-plugin-sdk/%s_os/%s_arch/%s",
-		version, autoscaling.Version, terraformVersion, meta.SDKVersionString(), runtime.GOOS, runtime.GOARCH) //nolint:staticcheck
-
-	client := &Client{
-		sdkClient: autoscaling.NewAPIClient(newAutoscalingConfig),
-	}
-	client.sdkClient.GetConfig().HTTPClient = &http.Client{Transport: utils.CreateTransport(insecure)}
+	config.MaxRetries = constant.MaxRetries
+	config.WaitTime = constant.MaxWaitTime
+	client := &Client{sdkClient: autoscaling.NewAPIClient(config)}
+	client.sdkClient.GetConfig().HTTPClient = &http.Client{Transport: shared.CreateTransport(clientOptions.SkipTLSVerify, clientOptions.Certificate)}
 
 	return client
 }