+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 ### Fixes
 - Remove cpu_family, availability_zone and rockylinux-8-GenericCloud-20230518 from docs
 - Do not return an error if `ionoscloud_object_storage_acesskey` is not found
+- Return early if dataplatform cluster is in `FAILED` state
+- `ionoscloud_s3_key` data source should require only `user_id`
 ## 6.7.2
 ### Fixes
 - Fix provider crashing when `canonical_user_id` is `nil` in the response for object storage access key